Rename parser package to skill for consistency with command
Rename internal/parser to internal/skill to mirror the command package
structure. This makes the codebase more consistent:
- skill package handles skill parsing (SKILL.md files)
- command package handles command parsing (*.md command files)

Also renamed local variable from 'skill' to 'parsedSkill' in main.go
to avoid shadowing the package name.
