+++ internal/skill/skill_test.go
@@ -1,4 +1,4 @@
-package parser
+package skill
 
 import (
 	"path/filepath"
